requirements.txt
# Core Processing
faster-whisper>=1.0.0
silero-vad>=5.0
rapidfuzz>=3.0.0
pydub>=0.25.0
soundfile>=0.12.0

# Video
moviepy>=1.0.3

# UI
PyQt6>=6.6.0
pyqtgraph>=0.13.0

# Utils
numpy>=1.24.0
Pillow>=10.0.0
